all product pages complete, admin panel and cart page remaining
