+++ client/src/app/products/leather/page.tsx
@@ -1,6 +1,7 @@
-import { Link } from "react-router-dom";
+import Image from "next/image";
+import Link from "next/link";
 
-export function LeatherProducts() {
+export default function LeatherProducts() {
   const products = [
     { name: "Premium Leather Bags", id: 1 },
     { name: "Handcrafted Wallets", id: 2 },
@@ -38,17 +39,19 @@ export function LeatherProducts() {
       <section className="relative h-screen overflow-hidden">
         {/* Background Image with Overlay */}
         <div className="absolute inset-0">
-          <div className="w-full h-full bg-gradient-to-b from-black/60 to-transparent">
-            <div className="w-full h-full bg-gray-300"></div>
+          <div className="w-full h-full">
+            <div className="w-full h-full bg-gradient-to-b from-neutral-600 to-transparent"></div>
           </div>
         </div>
 
         {/* Hero Content */}
         <div className="relative z-10 flex flex-col items-center justify-center h-full text-center px-4">
-          <h1 className="font-playfair font-bold text-4xl md:text-6xl lg:text-7xl text-veblyssSecondary mb-6 max-w-6xl">
+          <h1 className="font-playfair font-bold text-4xl md:text-6xl lg:text-7xl text-veblyssSecondary mb-6 max-w-6xl"
+            style={{ color: "#FFECE0", fontFamily: "Playfair Display" }}>
             Leather Products
           </h1>
-          <p className="font-opensans font-semibold text-xl md:text-2xl lg:text-3xl text-veblyssTextLight mb-12 max-w-4xl">
+          <p className="font-opensans font-semibold text-xl md:text-2xl lg:text-3xl text-veblyssTextLight mb-12 max-w-4xl"
+            style={{ color: "#FAF9F6", fontFamily: "Open Sans" }}>
             Crafted with care, tailored for global appeal.
           </p>
         </div>
@@ -58,7 +61,8 @@ export function LeatherProducts() {
       <section className="bg-veblyssSecondary py-16">
         <div className="container mx-auto px-4">
           <div className="text-center max-w-4xl mx-auto">
-            <h2 className="font-playfair font-normal text-4xl text-veblyssText mb-8">
+            <h2 className="font-playfair font-normal text-4xl text-veblyssText mb-8"
+              style={{ color: "#368581", fontFamily: "Playfair Display" }}>
               Premium Indian leather goods combining traditional craftsmanship
               with modern design, perfect for global markets seeking quality and
               authenticity.
@@ -70,47 +74,41 @@ export function LeatherProducts() {
       {/* Our Product Range */}
       <section className="bg-veblyssSecondary py-16">
         <div className="container mx-auto px-4">
-          <h2 className="font-playfair font-bold text-4xl lg:text-5xl text-veblyssPrimary text-center mb-12">
+          <h2
+            className="font-playfair font-bold text-4xl lg:text-5xl text-veblyssPrimary text-center mb-12"
+            style={{ color: "#368581", fontFamily: "Playfair Display" }}
+          >
             Our Product Range
           </h2>
-
-          {/* First Row - 3 Products */}
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-8 mb-12">
-            {products.slice(0, 3).map((product) => (
+      
+          {/* Dynamic Grid for Any Number of Products */}
+          <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-3 gap-8">
+            {products.map((product) => (
               <div
                 key={product.id}
-                className="bg-veblyssBackground rounded-xl shadow-lg overflow-hidden group hover:shadow-xl transition-all duration-300"
+                className="bg-white rounded-xl shadow-lg overflow-hidden group hover:shadow-xl transition-all duration-300"
               >
+                {/* Product Image */}
                 <div className="h-96 bg-gray-300 relative overflow-hidden">
-                  <div className="w-full h-full bg-gray-300"></div>
+                  <div className="w-full h-full bg-gray-300">
+                    <Image
+                      src='/images/placeholder.png'
+                      alt={product.name}
+                      className="object-cover w-full h-full"
+                      layout="fill"
+                    />
+                  </div>
                 </div>
+      
+                {/* Product Details */}
                 <div className="p-8 text-center">
                   <h3 className="font-playfair font-semibold text-2xl text-veblyssText mb-6">
                     {product.name}
                   </h3>
-                  <button className="bg-veblyssPrimary text-veblyssTextLight font-opensans font-bold text-lg px-8 py-3 rounded-xl hover:bg-opacity-90 transition-all duration-300">
-                    Check More
-                  </button>
-                </div>
-              </div>
-            ))}
-          </div>
-
-          {/* Second Row - 3 Products */}
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-            {products.slice(3, 6).map((product) => (
-              <div
-                key={product.id}
-                className="bg-veblyssBackground rounded-xl shadow-lg overflow-hidden group hover:shadow-xl transition-all duration-300"
-              >
-                <div className="h-96 bg-gray-300 relative overflow-hidden">
-                  <div className="w-full h-full bg-gray-300"></div>
-                </div>
-                <div className="p-8 text-center">
-                  <h3 className="font-playfair font-semibold text-2xl text-veblyssText mb-6">
-                    {product.name}
-                  </h3>
-                  <button className="bg-veblyssPrimary text-veblyssTextLight font-opensans font-bold text-lg px-8 py-3 rounded-xl hover:bg-opacity-90 transition-all duration-300">
+                  <button
+                    className="bg-veblyssPrimary text-veblyssTextLight font-opensans font-bold text-lg px-8 py-3 rounded-xl hover:bg-opacity-90 transition-all duration-300"
+                    style={{ backgroundColor: "#368581", color: "#FAF9F6" }}
+                  >
                     Check More
                   </button>
                 </div>
@@ -119,19 +117,22 @@ export function LeatherProducts() {
           </div>
         </div>
       </section>
+      
 
       {/* Why Choose Our Leather Products */}
       <section className="bg-veblyssSecondary py-16 relative">
         <div className="absolute inset-0 opacity-25">
-          <img
+          <Image
             src="https://api.builder.io/api/v1/image/assets/TEMP/24210383a5bf97fbabd00a6a6380b78f6db84648?width=2880"
             alt=""
             className="w-full h-full object-cover"
+            layout="fill"
           />
         </div>
 
         <div className="relative z-10 container mx-auto px-4">
-          <h2 className="font-playfair font-bold text-4xl lg:text-5xl text-veblyssPrimary text-center mb-12">
+          <h2 className="font-playfair font-bold text-4xl lg:text-5xl text-veblyssPrimary text-center mb-12 "
+            style={{color: "#368581", fontFamily: "Playfair Display"}}>
             Why Choose Our Leather Products
           </h2>
 
@@ -139,7 +140,7 @@ export function LeatherProducts() {
             {whyChooseFeatures.map((feature, index) => (
               <div
                 key={index}
-                className="bg-veblyssBackground rounded-xl px-8 py-6 shadow-lg hover:shadow-xl transition-all duration-300"
+                className="bg-white rounded-xl px-8 py-6 shadow-lg hover:shadow-xl transition-all duration-300"
               >
                 <h3 className="font-playfair font-bold text-xl text-veblyssText text-center">
                   {feature}
@@ -153,7 +154,8 @@ export function LeatherProducts() {
       {/* Ideal For */}
       <section className="bg-veblyssSecondary py-16">
         <div className="container mx-auto px-4">
-          <h2 className="font-playfair font-bold text-4xl lg:text-5xl text-veblyssPrimary text-center mb-12">
+          <h2 className="font-playfair font-bold text-4xl lg:text-5xl text-veblyssPrimary text-center mb-12"
+            style={{color: "#368581", fontFamily: "Playfair Display"}}>
             Ideal For
           </h2>
 
@@ -161,7 +163,7 @@ export function LeatherProducts() {
             {idealFor.map((item, index) => (
               <div
                 key={index}
-                className="bg-veblyssBackground rounded-xl px-8 py-6 shadow-lg hover:shadow-xl transition-all duration-300"
+                className="bg-white rounded-xl px-8 py-6 shadow-lg hover:shadow-xl transition-all duration-300"
               >
                 <h3 className="font-playfair font-bold text-xl text-veblyssText text-center">
                   {item}
@@ -179,18 +181,26 @@ export function LeatherProducts() {
             {relatedCategories.map((category) => (
               <div
                 key={category.name}
-                className="bg-veblyssBackground rounded-xl shadow-lg overflow-hidden group hover:shadow-xl transition-all duration-300"
+                className="bg-white rounded-xl shadow-lg overflow-hidden group hover:shadow-xl transition-all duration-300"
               >
                 <div className="h-64 bg-gray-300 relative overflow-hidden">
-                  <div className="w-full h-full bg-gray-300"></div>
+                  <div className="w-full h-full bg-gray-300">
+                    <Image
+                      src='/images/placeholder.png'
+                      alt={category.name}
+                      className="object-cover w-full h-full"
+                      layout="fill"
+                    />
+                  </div>
                 </div>
                 <div className="p-6 text-center">
                   <h3 className="font-playfair font-semibold text-lg text-veblyssText mb-4">
                     {category.name}
                   </h3>
                   <Link
-                    to={category.href}
-                    className="inline-block bg-veblyssPrimary text-veblyssTextLight font-opensans font-bold text-sm px-4 py-2 rounded-xl hover:bg-opacity-90 transition-all duration-300"
+                    href={category.href}
+                    className="inline-block text-veblyssTextLight font-opensans font-bold text-sm px-4 py-2 rounded-xl hover:bg-opacity-90 transition-all duration-300"
+                    style={{ backgroundColor: "#368581", color: "#FAF9F6" }}
                   >
                     Check More
                   </Link>
